pyproject.toml
[tool.poetry]
name = "kalada"
version = "0.0.1"
description = ""
authors = ["Alexander Lavrov <alexander.lavrov@elastoo.com>"]

[tool.poetry.dependencies]
python = "^3.8"
umongo = {extras = ["motor"], version = "^2.3.0"}
fastapi = "^0.61.2"
pydantic = "^1.7.2"
sitri = {extras = ["pyyaml"], version = "^0.7.7"}
dnspython = "^2.0.0"
httpx = "^0.16.1"
fake_headers = "^1.0.2"
pandas = "^1.1.4"
xmltodict = "^0.12.0"
loguru = "^0.5.3"
faust = "^1.10.4"
iso4217 = "^1.6.20180829"
ujson = "^4.0.1"
uvicorn = "^0.12.3"
python-dateutil = "^2.8.1"

[tool.poetry.dev-dependencies]
pre-commit = "^2.8.2"
ipython = "^7.19.0"

[tool.black]
line-length = 120
target-version = ['py38']
include = '\.pyi?$'
exclude = '''

(
  /(
      \.eggs         # exclude a few common directories in the
    | \.git          # root of the project
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
  )/
  | foo.py           # also separately exclude a file named foo.py in
                     # the root of the project
)
'''

[tool.vulture]
exclude = ["config/", "clients/", "wrappers/"]
min_confidence = 65
sort_by_size = true


[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
